+++ src/Component/Analytics/Facet/Pivot.php
@@ -99,6 +99,7 @@ class Pivot extends Configurable implements \JsonSerializable
         return $this;
     }
 
+    #[\ReturnTypeWillChange]
     /**
      * {@inheritdoc}
      */