+++ src/Component/Analytics/Facet/PivotFacet.php
@@ -67,6 +67,7 @@ class PivotFacet extends AbstractFacet
         return $this;
     }
 
+    #[\ReturnTypeWillChange]
     /**
      * {@inheritdoc}
      */