+++ src/Component/Analytics/Facet/RangeFacet.php
@@ -276,6 +276,7 @@ class RangeFacet extends AbstractFacet
         return $this;
     }
 
+    #[\ReturnTypeWillChange]
     /**
      * {@inheritdoc}
      */