+++ src/Component/Analytics/Grouping.php
@@ -134,6 +134,7 @@ class Grouping extends Configurable implements \JsonSerializable
         return $this;
     }
 
+    #[\ReturnTypeWillChange]
     /**
      * {@inheritdoc}
      */