+++ src/Core/Query/Result/Result.php
@@ -22,7 +22,7 @@ use Solarium\Exception\UnexpectedValueException;
  * This base class provides access to the response and decoded data. If you need more functionality
  * like resultset parsing use one of the subclasses
  */
-class Result implements ResultInterface
+class Result implements ResultInterface, \JsonSerializable
 {
     /**
      * Response object.
@@ -126,5 +126,14 @@ class Result implements ResultInterface
         }
 
         return $this->data;
+    }
+
+    #[\ReturnTypeWillChange]
+    /**
+     * {@inheritdoc}
+     */
+    public function jsonSerialize()
+    {
+        return $this->getData();
     }
 }