+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to the Solarium library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 ### Added
+- Results and Documents implement [JsonSerializable](https://www.php.net/manual/en/class.jsonserializable)
 
 ### Fixed
 - Adding nested child documents through `Document::setField()` and `Document::addField()`